+++ assets/enemies.toml
@@ -1,0 +1,19 @@
+#
+# @ Author: Niels Ouvrard - Diego Jiménez Ontiveros - Santiago Arreola Munguia
+# @ Create Time: 2024-11-13 13:27:13
+# @ Description:
+#
+
+[animations]
+frame_width = 16
+frame_height = 16
+
+[animations.idle]
+frame_count = 2
+y_position = 4
+x_positions = [0, 30]
+
+[animations.die]
+frame_count = 1
+y_position = 8
+x_positions = [46]
